requirements.txt
aws-cdk-lib
aiohttp==3.11.2
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.35.63
botocore==1.35.63
bs4==0.0.2
cachetools==5.5.0
cattrs==24.1.2
cdk-nag==2.34.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
clipboard==0.0.4
colorama==0.4.6
constructs==10.4.2
dataclasses-json==0.6.7
defusedxml==0.7.1
Events==0.5
faiss-cpu==1.9.0
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
jmespath==1.0.1
jsii==1.105.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-text-splitters==0.3.2
langsmith==0.1.143
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.13.5
numpy==1.26.4
opensearch-py==2.7.1
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pillow==11.0.0
portalocker==2.10.1
propcache==0.2.0
protobuf==5.28.3
publication==0.0.3
pyarrow==14.0.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
s3transfer==0.10.3
sacrebleu==2.4.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
streamlit==1.40.1
tabulate==0.9.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
typeguard==2.13.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.17.2
